+++ spring-boot-email-core/src/main/java/it/ozimov/springboot/templating/mail/service/ServiceStatus.java
@@ -1,0 +1,8 @@
+package it.ozimov.springboot.templating.mail.service;
+
+public enum ServiceStatus {
+
+    CREATED,
+    CLOSED
+
+}