+++ CHANGELOG.md
@@ -4,17 +4,36 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-## [0.2.0] - 2016-03-12
+
+## [0.3.0] - 2016-05-14
+### Added
+- Added support for **Mustache** (version from Spring Boot _1.3.4.RELEASE_)
+  and **Pebble** (_version 2.2.1_)
+
+### Changed
+- Main module changed to `spring-boot-email-core` to support generic template engines
+- Added specific modules based on known template engines (Freemarker, Mustache and Pebble)
+
+### Fixed
+- Closing mechanism of the scheduler is now responsive
+
+## [0.2.0] - 2016-05-12
 ### Added
 - Added implementation of a scheduler service with class `PriorityQyeyeSchedulerService`
+
+### Changed
 - Using Spring Boot _1.3.4.RELEASE_
 - Increased dependencies version
 - Removed dependency `zalando/problem-spring-web` (you can handle exceptions as you prefer)
 
 ## [0.1.1] - 2016-03-09
-### Added
+### Fixed
 - Minor bugfix for detecting mime type when not set for an email.
+
+### Changed
 - Increased dependencies version
+
+### Added
 - Added more unit test to increase coverage
 
 ## [0.1.0] - 2016-01-26